ADD: optimize a function of put tingCards together.
