+++ Mahjong/com/yuanfudao/megrez/app/codecamp/first/Logg.kt
@@ -10,3 +10,16 @@ object Logg {
         if (printCombine) println(s)
     }
 }
+
+fun <T> Result<T>.printError() {
+    this.onFailure {
+        if (it is DiscardException) {
+            Logg.debugLn("${it.msg}")
+        } else {
+            it.printStackTrace()
+        }
+    }
+}
+
+data class DiscardException(val msg: String? = "", val throws: Throwable? = null) :
+    java.lang.IllegalStateException(msg, throws)
